+++ src/App.tsx
@@ -1,0 +1,16 @@
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import HomePage from "@/pages/HomePage";
+import DesignSystemPage from "@/pages/_dev/design-system";
+
+function App() {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<HomePage />} />
+        <Route path="/_dev/design-system" element={<DesignSystemPage />} />
+      </Routes>
+    </BrowserRouter>
+  );
+}
+
+export default App;